refactorProjectStructure
