+++ src/components/common/delete.jsx
@@ -1,15 +1,16 @@
 import React from 'react'
 import { Button, Stack, Margin, Text } from '..'
-import { deleteData } from '../../common/services/firebase'
+import { deleteDocument } from '../../common/services/firebase'
 
-const Delete = ({ DataVolunteer, DataSkill, DataTrain, DataMember, DataHelp, DataEmergency, title }) => {
+const Delete = ({ DataVolunteer, DataSkill, DataTrain, DataMember, DataHelp, DataEmergency, title, setBar }) => {
+
     const DeleteDoc = () => {
-        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation', DataVolunteer)
-        if (title === 'УР ЧАДВАР') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/skills', DataSkill)
-        if (title === 'СУРГАЛТ') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining', DataTrain)
-        if (title === 'ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation', DataMember)
-        if (title === 'ТУСЛАМЖИЙН МЭДЭЭЛЭЛ') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation', DataHelp)
-        if (title === 'Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл') deleteData('/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson', DataEmergency)
+        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation/${DataVolunteer}`), setBar(false) }
+        if (title === `УР ЧАДВАР`) { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/skills/${DataSkill}`), setBar(false) }
+        if (title === `СУРГАЛТ`) { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining/${DataTrain}`), setBar(false) }
+        if (title === `ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ`) { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation/${DataMember}`), setBar(false) }
+        if (title === `ТУСЛАМЖИЙН МЭДЭЭЛЭЛ`) { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation/${DataHelp}`), setBar(false) }
+        if (title === `Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл`) { deleteDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson/${DataEmergency}`), setBar(false) }
     }
     return (<>
         <Text fontSize='12px' fontFamily='Roboto' lineHeight='12px' fontWeight='300'>
@@ -21,7 +22,7 @@ const Delete = ({ DataVolunteer, DataSkill, DataTrain, DataMember, DataHelp, Dat
                 <Margin size={[8, 0, 0, 0]}>
                     <Stack direction='row' justifyContent='space-around'>
                         <Button width='60px' height='23px' borderRadius='2px' bgColor='#33000000' bc='0.5px solid black'>Үгүй</Button>
-                        <Button width='60px' height='23px' borderRadius='2px' bgColor='#0066B3' color='white' onClick={DeleteDoc}> Тийм</Button>
+                        <Button width='60px' height='23px' borderRadius='2px' bgColor='#0066B3' color='white' onClick={DeleteDoc}>Тийм</Button>
                     </Stack>
                 </Margin>
             </Stack>
