+++ src/components/common/rightBarCheck.jsx
@@ -19,16 +19,26 @@ import Delete from './delete';
 import { addDocument, setDocument, useCollection, useDocument } from '../../common/services/firebase';
 import { Formik } from 'formik';
 import * as Yup from "yup";
+import { useEffect } from 'react';
 
 const RightBarCheck = ({ checkBar, setCheckBar, bar, setBar, title, el, headers, ind }) => {
-    const { data } = useDocument("/staticData/skillProps")
+    // const { loading, data } = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/skills');
+    // console.log(data[ind]?.uid)
+    // useEffect(() => {
+    //     console.log(loading)
+    //     if (!loading) {
+    //         console.log('asd')
+    //     }
+
+    // }, [])
+    const { data } = useDocument('/staticData/skillProps')
     const propsSelect = useMemo(() => {
-        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') return data.VolunteerNames;
-        if (title === 'СУРГАЛТ') return data.lessonNames
-        if (title === 'УР ЧАДВАР') return data.skillNames
-        if (title === 'ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ') return data.membershipNames;
-        if (title === 'ТУСЛАМЖИЙН МЭДЭЭЛЭЛ') return data.helpNames
-        if (title === 'Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл') return data.emergencyNames
+        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') return data?.VolunteerNames;
+        if (title === 'СУРГАЛТ') return data?.lessonNames
+        if (title === 'УР ЧАДВАР') return data?.skillNames
+        if (title === 'ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ') return data?.membershipNames;
+        if (title === 'ТУСЛАМЖИЙН МЭДЭЭЛЭЛ') return data?.helpNames
+        if (title === 'Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл') return data?.emergencyNames
     }, [data])
 
     const UptadeDoc = () => {
@@ -38,6 +48,7 @@ const RightBarCheck = ({ checkBar, setCheckBar, bar, setBar, title, el, headers,
     const New = () => {
         setCheckBar(false)
         setBar(true)
+        Add()
     }
     const [firstValue, setfirstValue] = useState('')
     const [secondValue, setsecondValue] = useState(el ? el[1] : '')
@@ -45,23 +56,40 @@ const RightBarCheck = ({ checkBar, setCheckBar, bar, setBar, title, el, headers,
     const [fourthValue, setfourthValue] = useState(el ? el[3] : '')
     const [fifthValue, setfifthValue] = useState(el ? el[4] : '')
 
-    const DataVolunteer = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation').dataId[ind]
-    const DataSkill = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/skills').dataId[ind]
-    const DataTrain = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/coveredTraining').dataId[ind]
-    const DataMember = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation').dataId[ind]
-    const DataHelp = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation').dataId[ind]
-    const DataEmergency = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson').dataId[ind]
-
-
+    const DataVolunteer = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation').data[ind]?.uid
+    const DataSkill = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/skills').data[ind]?.uid
+    const DataTrain = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/coveredTraining').data[ind]?.uid
+    const DataMember = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation').data[ind]?.uid
+    const DataHelp = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation').data[ind]?.uid
+    const DataEmergency = useCollection('/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson').data[ind]?.uid
     const [isSucces, setSucces] = useState('')
     const Add = () => {
-        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation', DataVolunteer, { additionalInformation: fourthValue, date: thirdValue, time: secondValue, volunteering: firstValue })
-            : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation', { additionalInformation: fourthValue, date: thirdValue, time: secondValue, volunteering: firstValue }, setSucces)
-        if (title === 'УР ЧАДВАР') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/skills', DataSkill, { skill: firstValue, skillLevel: secondValue, explanation: thirdValue }) : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/skills', { skill: firstValue, skillLevel: secondValue, explanation: thirdValue }, setSucces)
-        if (title === 'СУРГАЛТ') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining', DataTrain, { additionalInformation: fifthValue, trainingTime: fourthValue, trainingType: firstValue, when: secondValue, where: thirdValue }) : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining', { additionalInformation: fifthValue, trainingTime: fourthValue, trainingType: firstValue, when: secondValue, where: thirdValue }, setSucces)
-        if (title === 'ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation', DataMember, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, membershipType: firstValue }) : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation', { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, membershipType: firstValue }, setSucces)
-        if (title === 'ТУСЛАМЖИЙН МЭДЭЭЛЭЛ') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation', DataHelp, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, typeOfAssistance: firstValue }) : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation', { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, typeOfAssistance: firstValue }, setSucces)
-        if (title === 'Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл') el ? setDocument('/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson', DataEmergency, { information: firstValue, name: secondValue, phoneNumber: thirdValue }) : addDocument('/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson', { information: firstValue, name: secondValue, phoneNumber: thirdValue }, setSucces)
+        if (title === 'САЙН ДУРЫН АЖЛЫН МЭДЭЭЛЭЛ') {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation/${DataVolunteer}`, { additionalInformation: fourthValue, date: thirdValue, time: secondValue, volunteering: firstValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/volunteerWorkInformation`, { additionalInformation: fourthValue, date: thirdValue, time: secondValue, volunteering: firstValue })
+            setBar(false)
+        }
+        if (title === `УР ЧАДВАР`) {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/skills/${DataSkill}`, { skill: firstValue, skillLevel: secondValue, explanation: thirdValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/skills`, { skill: firstValue, skillLevel: secondValue, explanation: thirdValue })
+            setBar(false)
+        }
+        if (title === `СУРГАЛТ`) {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining/${DataTrain}`, { additionalInformation: fifthValue, trainingTime: fourthValue, trainingType: firstValue, when: secondValue, where: thirdValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/ coveredTraining`, { additionalInformation: fifthValue, trainingTime: fourthValue, trainingType: firstValue, when: secondValue, where: thirdValue })
+            setBar(false)
+        }
+
+        if (title === `ГИШҮҮНЧЛЭЛИЙН МЭДЭЭЛЭЛ`) {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation/${DataMember}`, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, membershipType: firstValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/membershipInformation`, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, membershipType: firstValue })
+            setBar(false)
+        }
+        if (title === `ТУСЛАМЖИЙН МЭДЭЭЛЭЛ`) {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation/${DataHelp}`, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, typeOfAssistance: firstValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/helpInformation`, { additionalInformation: fourthValue, endDate: thirdValue, startDate: secondValue, typeOfAssistance: firstValue })
+            setBar(false)
+        }
+        if (title === `Яаралтай үед холбоо барих гэр бүлийн гишүүний мэдээлэл`) {
+            el ? setDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson/${DataEmergency}`, { information: firstValue, name: secondValue, phoneNumber: thirdValue }) : addDocument(`/user/Y2Aiw9KPlijMFfTHIpsy/EmergencyContactPerson`, { information: firstValue, name: secondValue, phoneNumber: thirdValue }
+            )
+            setBar(false)
+        }
     }
 
 
@@ -79,10 +107,7 @@ const RightBarCheck = ({ checkBar, setCheckBar, bar, setBar, title, el, headers,
                                 <Padding size={[10, 50, 0, 50]}>
                                     <Formik
                                         initialValues={{ text: "" }}
-                                        onSubmit={async values => {
-                                            await new Promise(resolve => setTimeout(resolve, 500));
-                                            alert(JSON.stringify(values, null, 2));
-                                        }}
+
                                         validationSchema={Yup.object().shape({
                                             text: Yup.string()
                                                 .required("Required"),
@@ -224,7 +249,7 @@ const RightBarCheck = ({ checkBar, setCheckBar, bar, setBar, title, el, headers,
                                                                         {el ? <Stack direction='row' width='21vw' >
                                                                             <Image width='20px' height='20px' src={Trash} onClick={() => setDeleteDoc(!deleteDoc)} />
                                                                             <Position position='absolute' bottom='30px' left='-110px'>
-                                                                                {deleteDoc ? <Delete DataVolunteer={DataVolunteer} DataSkill={DataSkill} DataTrain={DataTrain} DataMember={DataMember} DataHelp={DataHelp} DataEmergency={DataEmergency} title={title} /> : ''}
+                                                                                {deleteDoc && <Delete DataVolunteer={DataVolunteer} DataSkill={DataSkill} DataTrain={DataTrain} DataMember={DataMember} DataHelp={DataHelp} DataEmergency={DataEmergency} title={title} ind={ind} setBar={setBar} />}
                                                                             </Position>
                                                                             <Margin size={[0, 0, 0, 10]}>
                                                                                 <Button width='60px' height='23px' bc='0.5px solid #00000033' color='black' bgColor='white' borderRadius='2px' onClick={() => setCheckBar(false)}>Болих</Button>
