+++ src/pages/admin/home.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
 
-import { useAuthContext } from '../common/context/AuthContext';
-import Admin from './admin';
-import { useCollection } from '../common/services/firebase';
+import { useAuthContext } from '../../common/context/AuthContext';
+import Admin from '.';
+import { useCollection } from '../../common/services/firebase';
 
 const AdminHome = () => {
   const { user } = useAuthContext();
@@ -12,14 +12,14 @@ const AdminHome = () => {
 
   useEffect(() => {
     if (user === null) {
-      router.push('/login');
+      router.push('/user');
     }
   }, [user]);
 
   userData.data.map((el) => {
     if (user.email === el.email) {
       if (el.admin === false) {
-        router.push('/userHome');
+        router.push('/user');
       }
     }
   });
