+++ src/pages/user/checker.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect } from 'react';
-import { useCollection } from '../common/services/firebase';
+import { useCollection } from '../../common/services/firebase';
 import { useRouter } from 'next/router';
-import { useAuthContext } from '../common/context/AuthContext';
-import { Loading } from '../components';
-import { LoaderContext, useLoaderContext } from '../common/context/LoaderContext';
+import { useAuthContext } from '../../common/context/AuthContext';
+import { Loading } from '../../components';
+import { LoaderContext, useLoaderContext } from '../../common/context/LoaderContext';
 
 const Checker = () => {
   const { setLoader } = useLoaderContext()
