+++ src/pages/user/register.jsx
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useRouter } from 'next/router';
-import { useAuthContext } from '..';
-import UserInfo from './UserInfo';
-import {Stepper} from '../components/common/stepper'
+import { useAuthContext } from '../../common/context/AuthContext';
+import UserInfo from './Info';
+import {Stepper} from '../../components/common/stepper'
 const data = {
   header: 'ХЭРЭГЛЭГЧИЙН ДЭЛГЭРЭНГҮЙ МЭДЭЭЛЭЛ БҮРТГЭХ',
   headers: ['ХУВЬ ХҮНИЙ ТАЛААРХ МЭДЭЭЛЭЛ', 'САЙН ДУРЫН АЖЛЫН ТУРШЛАГА', 'ХАМРАГДСАН СУРГАЛТ', 'УР ЧАДВАР'],
